# book-network/src/main/java/com/roque/book/book/BorrowedBookResponse.java
package com.roque.book.book;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class BorrowedBookResponse {
	private Integer id;
	private String title;
	private String authorName;
	private String isbn;
	private double rate;
	private boolean isReturned;
	private boolean isReturnApproved;

}