# book-network/src/main/java/com/roque/book/feedback/FeedbackResponse.java
package com.roque.book.feedback;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {
	private Double note;
	private String comment;
	private boolean ownFeedback;
}